feature: return errorMessage from useAuth.ts hook.
